# Conflicts:
#	src/apps/chem/test_SCFOperators.cc

+++ src/apps/chem/GuessFactory.h
@@ -30,7 +30,7 @@ std::vector<coord_3d> compute_centroids(const std::vector<Function<T,NDIM> > & v
 template<typename T, size_t NDIM>
 Vector<T,NDIM> tensor_to_coord(const Tensor<T>& t) {
 	Vector<T,NDIM> result;
-	MADNESS_ASSERT(t.size() >= NDIM);
+	MADNESS_ASSERT(size_t(t.size()) >= NDIM);
 	for (size_t i = 0; i < NDIM; ++i)
 		result[i] = t[i];
 	return result;
